+++ src/Viserio/Component/Container/LazyProxy/Instantiator/RealServiceInstantiator.php
@@ -1,11 +1,11 @@
 <?php
 declare(strict_types=1);
-namespace Viserio\Component\Container\Instantiator;
+namespace Viserio\Component\Container\LazyProxy\Instantiator;
 
 use Closure;
 use Psr\Container\ContainerInterface;
 use Viserio\Component\Contract\Container\Compiler\Definition as DefinitionContract;
-use Viserio\Component\Contract\Container\Instantiator as InstantiatorContract;
+use Viserio\Component\Contract\Container\LazyProxy\Instantiator as InstantiatorContract;
 
 class RealServiceInstantiator implements InstantiatorContract
 {
